+++ src/qt/rpcconsole.cpp
@@ -24,6 +24,7 @@
 #include <univalue.h>
 
 #ifdef ENABLE_WALLET
+#include <wallet/bdb.h>
 #include <wallet/db.h>
 #include <wallet/wallet.h>
 #endif
